feat:화상방 관련 기능 수정

+++ BackEnd/drinkus-api/src/main/java/com/ssafy/drinkus/room/service/RoomService.java
@@ -1,6 +1,7 @@
 package com.ssafy.drinkus.room.service;
 
 import com.ssafy.drinkus.common.NotFoundException;
+import com.ssafy.drinkus.common.NotMatchException;
 import com.ssafy.drinkus.interest.domain.Category;
 import com.ssafy.drinkus.interest.domain.CategoryRepository;
 import com.ssafy.drinkus.room.domain.Room;
@@ -12,6 +13,7 @@ import com.ssafy.drinkus.room.request.RoomUpdateRequest;
 import com.ssafy.drinkus.room.response.RoomInfoResponse;
 import com.ssafy.drinkus.room.response.RoomListResponse;
 import com.ssafy.drinkus.user.domain.User;
+import com.ssafy.drinkus.user.domain.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -19,9 +21,12 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Objects;
 import java.util.Optional;
 
 import static com.ssafy.drinkus.common.NotFoundException.CATEGORY_NOT_FOUND;
+import static com.ssafy.drinkus.common.NotFoundException.USER_NOT_FOUND;
+import static com.ssafy.drinkus.common.NotMatchException.USER_NOT_MATCH;
 
 @Slf4j
 @Service
@@ -31,12 +36,13 @@ public class RoomService {
     private final RoomRepository roomRepository;
     private final RoomQueryRepository roomQueryRepository;
     private final CategoryRepository categoryRepository;
+    private final UserRepository userRepository;
 
     //화상방 상세 조회
     public RoomInfoResponse findByRoomId(Long roomId){
         //아이디를 통해 화상방정보를 조회해온다
         Room room = roomRepository.findById(roomId)
-                .orElseThrow(() -> new NotFoundException(NotFoundException.USER_NOT_FOUND));
+                .orElseThrow(() -> new NotFoundException(NotFoundException.ROOM_NOT_FOUND));
         return RoomInfoResponse.from(room);
     }
 
@@ -55,12 +61,14 @@ public class RoomService {
 
     //화상방 생성
     @Transactional
-    public void createRoom(RoomCreateRequest request){
+    public void createRoom(User user, RoomCreateRequest request){
+        User findUser = userRepository.findById(user.getUserId())
+                .orElseThrow(() -> new NotFoundException(USER_NOT_FOUND));
         Category findCategory = categoryRepository.findById(request.getCategory().getCategoryId())
                 .orElseThrow(() -> new NotFoundException(CATEGORY_NOT_FOUND));
         Room room = Room.createRoom(
                 request.getRoomName(),
-                request.getRoomAdminId(),
+                findUser,
                 request.getRoomPw(),
                 request.getPlaceTheme(),
                 request.getPeopleLimit(),
@@ -78,12 +86,16 @@ public class RoomService {
 
     //화상방 수정
     @Transactional
-    public void updateRoom(Long roomId, RoomUpdateRequest request){
+    public void updateRoom(User user, Long roomId, RoomUpdateRequest request){
         Room findroom = roomRepository.findByRoomId(roomId)
                 .orElseThrow(() -> new NotFoundException(NotFoundException.ROOM_NOT_FOUND));
 
         Category findCategory = categoryRepository.findById(request.getCategory().getCategoryId())
                 .orElseThrow(() -> new NotFoundException(CATEGORY_NOT_FOUND));
+
+        if(!user.getUserId().equals(findroom.getUser().getUserId())){
+            throw new NotMatchException(USER_NOT_MATCH);
+        }
 
         findroom.updateRoom(
                 request.getRoomName(),
@@ -101,11 +113,12 @@ public class RoomService {
 
     @Transactional
     //화상방 삭제
-    public void deleteRoom(Long roomId){
+    public void deleteRoom(User user, Long roomId){
         Room findroom = roomRepository.findByRoomId(roomId)
                 .orElseThrow(() -> new NotFoundException(NotFoundException.ROOM_NOT_FOUND));
-
+        if(!user.getUserId().equals(findroom.getUser().getUserId())){
+            throw new NotMatchException(USER_NOT_MATCH);
+        }
         roomRepository.deleteById(roomId);
     }
-
 }
